# internal/docs/openapi.yaml
openapi: 3.1.0
info:
  title: socialstats API
  version: 1.0.0
  summary: Public view counts for a social video URL.
  description: |
    Give the service a video URL and it returns that video's public metrics.

    YouTube, TikTok and Meta (Facebook + Instagram) are implemented.

    **Reliability differs by platform.** YouTube is read through the official
    Data API and is reliable. Neither TikTok nor Meta offers an API for
    arbitrary video URLs, so their counters come from the public pages.

    - TikTok serves an anti-bot challenge to a substantial share of requests,
      which the service retries. Expect occasional `502 upstream_blocked` on
      TikTok URLs, and retry — it usually succeeds.
    - Instagram is read from its login-free embed render and is reasonably
      reliable. Facebook prefers the Graph API when a token is configured and
      can see the object, and otherwise reads the public page — for which
      Facebook often serves a login wall instead. That is also
      `502 upstream_blocked`, but retrying it will not help.

    **There is a dashboard too.** The same binary serves a server-rendered UI at
    `/`, sharing these endpoints' service layer, session cookies and CSRF token.
    This document covers the JSON API only; the HTML pages are not part of the
    contract and may change freely.

    **Two ways to authenticate, and the choice matters.** A session cookie or an
...
    second. Cookie-authenticated writes must echo the CSRF token; bearer and
...
servers:
...
paths:
...
  /v1/videos:
    get:
      tags: [tracking]
      summary: The caller's tracked videos
      description: |
        Returns the caller's list with each video's current counters and its
        growth over the requested window.

        `views_gained` is the current view count minus the last reading at or
        before the window's start. It is **absent rather than zero** when there
        is no baseline — a video added yesterday has no week-old reading — and
        it is **signed**, because platforms revise view counts downward and a
        negative is a real measurement.
      operationId: listVideos
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: window
          in: query
          description: How far back the growth baseline is taken from. Max one year.
          schema: { type: string, default: 168h, examples: ['24h', '168h'] }
        - name: platform
          in: query
          schema: { $ref: '#/components/schemas/Platform' }
        - name: sort
          in: query
          schema:
            type: string
            enum: [views, gained, recent, title, fetched]
            default: views
        - name: sparkline
          in: query
          description: |
            Attach this many recent readings to each entry, oldest first. Off by
            default: it costs one query per video, which is fine for a page and
            wasteful for a client that only wants the numbers.
          schema: { type: integer, minimum: 0, maximum: 200, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: The caller's tracked videos.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VideoListEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    post:
      tags: [tracking]
      summary: Track a video
      description: |
        Resolves the URL, records the video and starts tracking it for the
        caller — then **fetches it immediately**, so the response carries real
        numbers rather than an empty row that fills in within the next interval.

        A failed first fetch still tracks the video and still returns `201`. The
        failure is recorded on the row (`schedule.last_fetch_status`) and in the
        fetch log, and the poller retries: a TikTok challenge at 14:03 is not a
        reason to refuse to track a video.

        The exception is short links (`vm.tiktok.com`, `fb.watch`,
        `/share/...`), which carry no id until the redirect is followed. A
        failed fetch there means the video genuinely cannot be identified, and
        the request fails rather than tracking something unknown.
      operationId: trackVideo
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TrackRequest' }
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/TrackRequest' }
      responses:
        '201':
          description: |
            Now tracked. Whether the first fetch succeeded is reported by the
            entry's own fetch status, not by the status code.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/LimitReached' }
        '502': { $ref: '#/components/responses/UpstreamError' }
...
  /v1/videos/{id}/refresh:
    post:
      tags: [tracking]
      summary: Bring the next fetch forward
      description: |
        Schedules the video to be fetched on the next poll tick, for a caller
        who does not want to wait for the interval.

        It returns `202` because it queues rather than fetches. Fetching inline
        on demand would let one impatient account spend the whole deployment's
        TikTok and Meta budget, and the poller is already what paces platform
        access.
      operationId: refreshVideo
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/VideoID'
      responses:
        '202':
          description: Queued for the next poll tick.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    properties:
                      queued: { type: boolean }
                      video: { $ref: '#/components/schemas/Entry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotTracked' }
        '410':
          description: |
            The video has been retired: the platform said it no longer exists,
            repeatedly. Its history is still readable; it is simply not fetched
            any more.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /v1/auth/logout-all:
    post:
      tags: [auth]
      summary: End every session for this account
      description: |
        Revokes every session the account has, on every device.

        This is what server-side sessions buy that a stateless signed token
        would not: revocation is a DELETE, not a denylist that has to be
        consulted on every request until the token would have expired anyway.
      operationId: logoutEverywhere
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        '200':
          description: Every session was revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedOutEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/auth/password:
    post:
      tags: [auth]
      summary: Change password
      description: |
        Verifies the current password and replaces it.

        Every session is revoked, including the one making the request. The
        usual reason to change a password is believing someone else has it, and
        leaving their session signed in would make the change ceremonial.
      operationId: changePassword
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '200':
          description: Password changed and every session revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedOutEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ss_session
      description: |
        The session cookie set by register and login. HttpOnly, SameSite=Lax and
        Secure outside development.

        Requests authenticated this way must also pass the CSRF check on unsafe
        methods: echo the readable CSRF cookie back in `X-CSRF-Token` or a
        `csrf_token` form field.
    bearerToken:
      type: http
      scheme: bearer
      description: |
        The same session token, sent as `Authorization: Bearer <token>`. Login
        returns it in the body for exactly this purpose.

        Bearer requests are exempt from CSRF, and not as a convenience: CSRF
        exists because browsers attach cookies to cross-site requests
        automatically. They do not attach this header on a third party's behalf,
        so the request cannot be forged.

        A request with an `application/json` body is exempt for the same kind of
        reason: an HTML form can only send urlencoded, multipart or text/plain,
        so a JSON content type is itself proof the request did not come from
        one. Every form-encoded write is checked, including sign-in and sign-up.
  parameters:
    VideoID:
      name: id
      in: path
      required: true
      description: |
        The video's public uuid. Internal row ids are never exposed, so the ids
        in these URLs cannot be enumerated.
      schema: { type: string, format: uuid }
    RequestID:
      name: X-Request-Id
      in: header
      required: false
      description: |
        Correlation id. When supplied it is echoed back and used in logs;
        otherwise the service generates one.
      schema:
        type: string
  headers:
    XRequestID:
      description: The request's correlation id, echoed on every response.
      schema: { type: string }
...
  schemas:
    Platform:
      type: string
      description: A supported social video platform.
      enum: [youtube, meta, tiktok]
...
    TrackRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Any URL form the providers accept; see the /v1/stats parameter.
          examples: ['https://www.tiktok.com/@user/video/7249376077976472833']
        label:
          type: string
          description: The caller's own name for this video. Per-user, not shared.
          examples: ['Q3 launch']
...
    Entry:
      type: object
      required: [video, added_at, fresh]
      properties:
        video: { $ref: '#/components/schemas/TrackedVideo' }
        label: { type: string }
        added_at: { type: string, format: date-time }
        views_gained:
          type: integer
          format: int64
          description: |
            Current view count minus the baseline. Absent when there is no
            baseline in the window, and signed, because counters are not
            monotonic.
        baseline_at:
          type: string
          format: date-time
          description: |
            When the reading `views_gained` was measured against was taken.
            Without it the delta is uninterpretable: "+1,200 since some
            unspecified point" is not a fact anyone can use.
        sparkline:
          type: array
          description: Recent readings, oldest first. Only when `sparkline` was requested.
          items: { $ref: '#/components/schemas/Snapshot' }
        fresh:
          type: boolean
          description: |
            Whether the latest reading is recent enough to show without a
            caveat, measured against this video's own fetch interval rather than
            a fixed cutoff.
    EntryEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Entry' }
    VideoListEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [videos, count, window, limit, offset]
          properties:
            videos:
              type: array
              items: { $ref: '#/components/schemas/Entry' }
            count: { type: integer }
            window: { type: string }
            limit: { type: integer }
            offset: { type: integer }
...
    SignedOutEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [signed_out]
          properties:
            signed_out: { type: boolean }
            sessions_revoked: { type: integer, format: int64 }
            password_changed: { type: boolean }
...
    Error:
      type: object
      description: The single error shape returned by every endpoint.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message:
              type: string
              description: Human-readable explanation. Do not parse.
            request_id:
              type: string
              description: Correlation id for this request; quote it in bug reports.
  responses:
...
    Unauthorized:
      description: |
        Not signed in, or the credentials were wrong.

        `invalid_credentials` is returned for an unknown address, a wrong
        password and a suspended account alike, so the endpoint cannot be used
        to learn which of the three it was.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            badCredentials:
              value:
                error:
                  code: invalid_credentials
                  message: invalid email or password
                  request_id: 4fba7d8d0838ef557fff84af4bd6d83d
            notSignedIn:
              value:
                error:
                  code: unauthenticated
                  message: sign in to use this endpoint
                  request_id: 0d8baf20a202350408789ab8e33b53b8
    Forbidden:
      description: |
        The request was understood and refused. Either the CSRF check failed on
        a cookie-authenticated write, or registration is closed here.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            csrf:
              value:
                error:
                  code: csrf_failed
                  message: the csrf token was missing or did not match
                  request_id: c034335474cab0b94f7837a7eb56c38f
            closed:
              value:
                error:
                  code: registration_closed
                  message: registration is closed on this deployment
                  request_id: 8a1f2c9d4e6b70318a5c2d9e7f4b1a63
...
    NotTracked:
      description: |
        No such video on the caller's list.

        A video the caller does not track returns `404` rather than `403`, so
        the endpoint does not confirm that a video other people are tracking
        exists.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            notTracked:
              value:
                error:
                  code: not_found
                  message: no such record
                  request_id: 6f1c2a9d4e6b70318a5c2d9e7f4b1a63
    LimitReached:
      description: |
        The account is at its tracking limit.

        Tracking is what creates polling work, and polling work is spent against
        upstream quota and anti-bot budgets that everyone on the deployment
        shares — so one account cannot take all of it.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            full:
              value:
                error:
                  code: limit_reached
                  message: 'account limit reached: you are tracking 200 videos, the limit is 200'
                  request_id: 91a0bcde2f3147a5b8c6d2e9f0a1b234
...
    BadRequest:
      description: The request was malformed, or the URL is not a supported video URL.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            missingURL:
              summary: No url supplied
              value:
                error:
                  code: missing_url
                  message: query parameter 'url' is required
                  request_id: 09f67ff774aefae24185e16d344b930b
            invalidURL:
              summary: Not a parseable video URL
              value:
                error:
                  code: invalid_url
                  message: the supplied url could not be parsed as a supported video url
                  request_id: 09f67ff774aefae24185e16d344b930b
            unsupportedPlatform:
              summary: A host no provider claims
              value:
                error:
                  code: unsupported_platform
                  message: no provider handles that url
                  request_id: add2d0a65dc565fa18585bb50d243bc3
...
    UpstreamError:
      description: |
        The platform failed this request. Two codes share this status:

        - `upstream_error` — the platform's API returned an error the service
          could not classify further.
        - `upstream_blocked` — the platform served an anti-bot challenge or a
          login wall instead of content, and the service exhausted its retries.
          It applies to the platforms read from public pages rather than an API.
          On TikTok this is a random challenge and retrying usually succeeds; on
          Facebook it is usually a login wall, which retrying will not clear.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            upstream:
              summary: The platform's API errored
              value:
                error:
                  code: upstream_error
                  message: the upstream platform returned an error
                  request_id: c1cbd7913f833ec7d25b8e5dcedb7b67
            blocked:
              summary: TikTok served an anti-bot challenge; retry
              value:
                error:
                  code: upstream_blocked
                  message: the platform served an anti-bot challenge instead of content; retry later
                  request_id: 2b3d028dfea63146616e5ccc4d398546
            loginWall:
              summary: Facebook served a login wall; retrying will not help
              value:
                error:
                  code: upstream_blocked
                  message: the platform served an anti-bot challenge instead of content; retry later
                  request_id: 9f0a1c7de2b34ac59a1e6b8d4f2c7a10
...